+++ rest/src/main/java/org/n52/web/ctrl/CategoriesParameterController.java
@@ -29,6 +29,7 @@
 package org.n52.web.ctrl;
 
 
+import org.n52.io.request.IoParameters;
 import org.n52.io.response.CategoryOutput;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -37,4 +38,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(UrlSettings.COLLECTION_CATEGORIES)
 public class CategoriesParameterController extends ParameterRequestMappingAdapter<CategoryOutput> {
 
+    @Override
+    public String getCollectionPath(String hrefBase) {
+        UrlHelper urlhelper = new UrlHelper();
+        return urlhelper.constructHref(hrefBase, UrlSettings.COLLECTION_CATEGORIES);
+    }
+
+    @Override
+    protected int getElementCount(IoParameters queryMap) {
+        return super.getEntityCounter().getCategoryCount(IoParameters.ensureBackwardsCompatibility(queryMap));
+    }
 }