+++ rest/src/main/java/org/n52/web/ctrl/GeometriesController.java
@@ -28,6 +28,7 @@
  */
 package org.n52.web.ctrl;
 
+import org.n52.io.request.IoParameters;
 import org.n52.io.response.GeometryInfo;
 import org.n52.series.spi.geo.TransformingGeometryOutputService;
 import org.n52.series.spi.srv.ParameterService;
@@ -41,6 +42,18 @@ public class GeometriesController extends ParameterRequestMappingAdapter<Geometr
     @Override
     public void setParameterService(ParameterService<GeometryInfo> parameterService) {
         super.setParameterService(new TransformingGeometryOutputService(parameterService));
+    }
+
+    @Override
+    public String getCollectionPath(String hrefBase) {
+        UrlHelper urlhelper = new UrlHelper();
+        return urlhelper.constructHref(hrefBase, UrlSettings.COLLECTION_GEOMETRIES);
+    }
+
+    @Override
+    protected int getElementCount(IoParameters queryMap) {
+        //TODO(specki): Implementation of getCount  for Geometries in EntityCounter
+        return -1;
     }
 
 }