+++ rest/src/main/java/org/n52/web/ctrl/PlatformsParameterController.java
@@ -28,6 +28,7 @@
  */
 package org.n52.web.ctrl;
 
+import org.n52.io.request.IoParameters;
 import org.n52.io.response.PlatformOutput;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -36,4 +37,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(path = UrlSettings.COLLECTION_PLATFORMS)
 public class PlatformsParameterController extends ParameterRequestMappingAdapter<PlatformOutput> {
 
+    @Override
+    public String getCollectionPath(String hrefBase) {
+        UrlHelper urlhelper = new UrlHelper();
+        return urlhelper.constructHref(hrefBase, UrlSettings.COLLECTION_PLATFORMS);
+    }
+
+    @Override
+    protected int getElementCount(IoParameters queryMap) {
+        return super.getEntityCounter().getPlatformCount(queryMap);
+    }
 }