+++ rest/src/main/java/org/n52/web/ctrl/TimeseriesMetadataController.java
@@ -28,6 +28,7 @@
  */
 package org.n52.web.ctrl;
 
+import org.n52.io.request.IoParameters;
 import org.n52.io.response.dataset.TimeseriesMetadataOutput;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -37,5 +38,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(value = UrlSettings.COLLECTION_TIMESERIES)
 public class TimeseriesMetadataController extends ParameterRequestMappingAdapter<TimeseriesMetadataOutput> {
 
+    @Override
+    public String getCollectionPath(String hrefBase) {
+        UrlHelper urlhelper = new UrlHelper();
+        return urlhelper.constructHref(hrefBase, UrlSettings.COLLECTION_TIMESERIES);
+    }
+
+    @Override
+    protected int getElementCount(IoParameters queryMap) {
+        return super.getEntityCounter().getTimeseriesCount();
+    }
     // resource controller for timeseries metadata
 }